+++ kernel/src/planificacionUtils.c
@@ -77,10 +77,10 @@ op_code esperar_respuesta_memoria(int socket_memoria) {
  	return tiempo_actual;
 }*/
 
-long long get_time() {
+uint32_t get_time() {
     struct timeval te;
     gettimeofday(&te, NULL); // get current time
-    long long milliseconds = te.tv_sec*1000LL + te.tv_usec/1000; // calculate milliseconds
+    uint32_t milliseconds = te.tv_sec*1000LL + te.tv_usec/1000; // calculate milliseconds
     // printf("milliseconds: %lld\n", milliseconds);
     return milliseconds;
 }
@@ -231,9 +231,10 @@ void transicion_interrupcion(){
 		interrupcion = 1;
 		pthread_mutex_unlock(&mutex_interrupcion);
 		interrumpir_cpu();
-	} else {
+	}
+	else {
 		pthread_mutex_unlock(&mutex_exec);
-				 	}
+	}
 }
